% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bllnn-package.R
\docType{package}
\name{bllnn-package}
\alias{bllnn}
\alias{bllnn-package}
\title{bllnn: Bayesian Last Layer Neural Networks as Composable Gibbs Blocks}
\description{
Exposes a Bayesian Last Layer neural network as a composable conditional-model block for user-written Gibbs samplers, in the spirit of 'dbarts' for Bayesian Additive Regression Trees. The caller supplies a residual and a noise level and receives one honest posterior draw of the nonlinear function. Intended for partial-linear causal inference where the linear coefficient requires an honest credible interval.
}
\author{
\strong{Maintainer}: Saptarshi Niyogi \email{suvo123123@gmail.com} [copyright holder]

Authors:
\itemize{
  \item Saptarshi Niyogi \email{suvo123123@gmail.com} [copyright holder]
  \item Saptarshi Chakraborty
}

}
\keyword{internal}
