% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyagamma.R
\name{pg_mean}
\alias{pg_mean}
\title{Mean of PG(1, z)}
\usage{
pg_mean(z)
}
\arguments{
\item{z}{Tilting parameter.}
}
\value{
\verb{tanh(z / 2) / (2 z)}, with the limit \code{1/4} taken at zero.
}
\description{
Closed form, used to check the sampler and available for diagnostics.
}
\examples{
pg_mean(c(0, 1, 4))

}
