% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyagamma.R
\name{pg_var}
\alias{pg_var}
\title{Variance of PG(1, z)}
\usage{
pg_var(z)
}
\arguments{
\item{z}{Tilting parameter.}
}
\value{
The variance, with the limit \code{1/24} taken at zero.
}
\description{
Variance of PG(1, z)
}
\examples{
pg_var(c(0, 1, 4))

}
